requirements.txt
crawlee[playwright]>=0.3.0
playwright>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.5.0
jinja2>=3.1.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
unidecode>=1.3.7
Pillow>=10.1.0
reportlab>=4.0.7

